+++ BLL/PropertyClasses/Transaction/PurchaseReturn_Property.cs
@@ -12,7 +12,7 @@ namespace BLL.PropertyClasses.Transaction
         public string return_date { get; set; }
         public Int64 firm_id { get; set; }
         public Int64 gst_id { get; set; }
-        public Int64 total_pcs { get; set; }
+        public decimal total_pcs { get; set; }
         public decimal gross_amount { get; set; }
         public string purchase_bill_no { get; set; }
         public decimal cgst_rate { get; set; }
@@ -42,13 +42,13 @@ namespace BLL.PropertyClasses.Transaction
         public Int64 color_id { get; set; }
         public Int64 size_id { get; set; }
         public Int64 unit_id { get; set; }
-        public int pcs { get; set; }
+        public decimal pcs { get; set; }
         public decimal rate { get; set; }
         public decimal amount { get; set; }
         public decimal discount { get; set; }
         public Int64 voucher_no { get; set; }
         public string remarks { get; set; }
-        public int old_pcs { get; set; }
+        public decimal old_pcs { get; set; }
         public int flag { get; set; }
         public Int64 old_item_id { get; set; }
         public Int64 old_color_id { get; set; }
